Factor out remaining functions from cli/utils/__init__.py
At this point we may as well finish the job and make this file a
namespace only.
