+++ meson.build
@@ -22,7 +22,8 @@ py_src = {
                    'optimise_dipole_parameter.py', 'powder_map.py',
                    'show_sampling.py'],
   'euphonic/cli/utils': [
-      '__init__.py', '_band_structure.py', '_cli_parser.py', '_grids.py', '_loaders.py', '_pdos.py'
+      '__init__.py', '_band_structure.py', '_cli_parser.py', '_dw.py',
+      '_grids.py', '_kwargs.py', '_loaders.py', '_pdos.py', '_plotting.py'
   ],
   'euphonic/data': ['__init__.py', 'bluebook.json', 'sears-1992.json',
                     'reciprocal_spectroscopy_definitions.txt'],
